Boundaries working. Some testing. Some refactoring.

+++ vhl/src/main/java/info/cinow/repository/CensusTractDao.java
@@ -1,6 +1,5 @@
 package info.cinow.repository;
 
-import org.geojson.FeatureCollection;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
@@ -10,7 +9,7 @@ import info.cinow.model.CensusTract;
 /**
  * CensusTractDao
  */
-public interface CensusTractDao extends CrudRepository<CensusTract, Long> {
+public interface CensusTractDao extends CrudRepository<CensusTract, Integer> {
 
     @Query(value = "SELECT gid FROM (SELECT gid,  ST_CONTAINS(GEOM, ST_MakePoint(:longitude, :latitude)) AS CONTAINING FROM bexartracts_2010) ss WHERE ss.CONTAINING = true", nativeQuery = true)
     public Long getContainingTract(@Param("longitude") double longitude, @Param("latitude") double latitude);
